requirements.txt
pyautogen==0.3.0
uvicorn==0.30.6
scikit-learn==1.5.2
fastapi==0.115.0
xgboost==2.1.1
sse-starlette==2.1.3
pydantic==2.9.2